refactor: 트리밍 스케줄러 별도 클래스 분리 및 예외 처리 추가
